update：物体生成初步
打了个底，还需要写：
1.客户端物体生成
2.per Client的ds id构造->snapshot内存构造

+++ Assets/StargateNet/StargateNet/Base/SgNetworkUtil.cs
@@ -6,7 +6,7 @@ namespace StargateNet
     public static class SgNetworkUtil
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static long AlignTo(long value, long alignment)
+        public static int AlignTo(int value, int alignment)
         {
             // (x + (align - 1)) & ~(align - 1)
             return (value + (alignment - 1)) & ~(alignment - 1);
